# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hybrid firewall in Go using public blocklists",
        "title": "Meds: net healing",
        "contact": {
            "name": "cnaize",
            "url": "https://github.com/cnaize"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1.2.0"
    },
    "paths": {
        "/v1/asns/blocklist/exclude": {
            "get": {
                "description": "get all excluded ASNs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Get excluded ASNs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetASNsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert ASNs to excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Upsert excluded ASNs",
                "parameters": [
                    {
                        "description": "ASNs to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertASNsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove ASNs from excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Remove excluded ASNs",
                "parameters": [
                    {
                        "description": "ASNs to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveASNsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/asns/blocklist/exclude/{asn}": {
            "get": {
                "description": "check if an ASN is excluded",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Check excluded ASN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ASN to check",
                        "name": "asn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckASNResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/v1/asns/blocklist/include": {
            "get": {
                "description": "get all included ASNs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Get included ASNs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetASNsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert ASNs to includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Upsert included ASNs",
                "parameters": [
                    {
                        "description": "ASNs to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertASNsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove ASNs from includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Remove included ASNs",
                "parameters": [
                    {
                        "description": "ASNs to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveASNsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/asns/blocklist/include/{asn}": {
            "get": {
                "description": "check if an ASN is included",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ASNs"
                ],
                "summary": "Check included ASN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ASN to check",
                        "name": "asn",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckASNResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/v1/countries/blocklist": {
            "get": {
                "description": "get all blocked countries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Get blocked countries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetCountriesResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert countries to blocklist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Upsert blocked countries",
                "parameters": [
                    {
                        "description": "countries to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertCountriesReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove countries from blocklist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Remove blocked countries",
                "parameters": [
                    {
                        "description": "countries to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveCountriesReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/countries/blocklist/{country}": {
            "get": {
                "description": "check if a country is blocked",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Countries"
                ],
                "summary": "Check blocked country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "country to check",
                        "name": "country",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckCountryResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.CheckASNResp": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                }
            }
        },
        "api.CheckCountryResp": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                }
            }
        },
...
        "api.GetASNsResp": {
            "type": "object",
            "properties": {
                "asns": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        206831,
                        400328
                    ]
                }
            }
        },
        "api.GetCountriesResp": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "fr",
                        "de"
                    ]
                }
            }
        },
...
        "api.RemoveASNsReq": {
            "type": "object",
            "properties": {
                "asns": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        206831,
                        400328
                    ]
                }
            }
        },
        "api.RemoveCountriesReq": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "fr",
                        "de"
                    ]
                }
            }
        },
...
        "api.UpsertASNsReq": {
            "type": "object",
            "properties": {
                "asns": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        206831,
                        400328
                    ]
                }
            }
        },
        "api.UpsertCountriesReq": {
            "type": "object",
            "properties": {
                "countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "fr",
                        "de"
                    ]
                }
            }
        },
...
    }
}